requirements.txt
fuse-python
boto3
xdg
